# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API LiqaaSpace",
        "description": "Documentation de l'API LiqaaSpace",
        "contact": {
            "name": "Équipe LiqaaSpace",
            "email": "contact@liqaaspace.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Liste toutes les réservations",
                "operationId": "26909e6f6eb4a49ab5428889a7ff1067",
                "responses": {
                    "200": {
                        "description": "Liste des réservations récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Créer une nouvelle réservation",
                "operationId": "b1466042c18abd23d64464649f2442c3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "salle_id",
                                    "user_id",
                                    "date_debut",
                                    "date_fin",
                                    "statut"
                                ],
                                "properties": {
                                    "salle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-05-22 10:00:00"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-05-22 12:00:00"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "confirmé",
                                            "annulé"
                                        ],
                                        "example": "confirmé"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Réservation créée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Afficher une réservation spécifique",
                "operationId": "f896182177203ee52325e4cbee5a7c7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la réservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation récupérée avec succès"
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Mettre à jour une réservation",
                "operationId": "1564fc30d0991a49c920d815c32c8a61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la réservation à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "salle_id",
                                    "user_id",
                                    "date_debut",
                                    "date_fin",
                                    "statut"
                                ],
                                "properties": {
                                    "salle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-05-22 10:00:00"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-05-22 12:00:00"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "confirmé",
                                            "annulé"
                                        ],
                                        "example": "confirmé"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réservation mise à jour avec succès"
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Réservations"
                ],
                "summary": "Supprimer une réservation",
                "operationId": "c46682105a40a6b2324956601b05bbf4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la réservation à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réservation supprimée avec succès"
                    },
                    "404": {
                        "description": "Réservation non trouvée"
                    }
                }
            }
        },
...
    },
...
}